# Improve perf in discovery jobs metrics to data lookup