--- a/Makefile
+++ b/Makefile
@@ -9,7 +9,7 @@ build:
 	go build -v -ldflags "$(GO_LDFLAGS)" -o yace ./cmd/yace
 
 test:
-	go test -v -race -count=1 ./...
+	go test -v -bench=^$ -race -count=1 ./...
 
 lint:
 	golangci-lint run -v -c .golangci.yml