This PR addessed some performance issues we've been seeing with running YACE discovery with large enough jobs. We noticed there was a lot of cpu % usage around `findGetMetricDataByID`. This lead us to believe this is a hot path, and as well, performs lots of side-allocations.

This PR re-implements that logic with a map to lookup the corresponding metric data, rather than looping over. Also, a benchmark is added for further development.

The benchstat below shows an improvement of around ~90% in both cpu and memory. 

benchstat
```
goos: darwin
goarch: arm64
pkg: github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/job
                                      │    main.p     │                pr.p                │
                                      │    sec/op     │   sec/op     vs base               │
MapResultsToMetricDatas/small_case-10    271.13µ ± 3%   10.35µ ± 1%  -96.18% (p=0.002 n=6)
MapResultsToMetricDatas/big_case-10     28916.8µ ± 4%   927.1µ ± 6%  -96.79% (p=0.002 n=6)
geomean                                   2.800m        97.96µ       -96.50%

                                      │     main.p      │                pr.p                 │
                                      │      B/op       │     B/op      vs base               │
MapResultsToMetricDatas/small_case-10     364367.5 ± 0%     444.0 ± 0%  -99.88% (p=0.002 n=6)
MapResultsToMetricDatas/big_case-10     36133.26Ki ± 0%   15.32Ki ± 2%  -99.96% (p=0.002 n=6)
geomean                                    3.502Mi        2.577Ki       -99.93%

                                      │    main.p     │               pr.p                │
                                      │   allocs/op   │ allocs/op   vs base               │
MapResultsToMetricDatas/small_case-10   7921.000 ± 0%   2.000 ± 0%  -99.97% (p=0.002 n=6)
MapResultsToMetricDatas/big_case-10     809687.5 ± 0%   356.0 ± 3%  -99.96% (p=0.002 n=6)
geomean                                   80.08k        26.68       -99.97
```

main
```
goos: darwin
goarch: arm64
pkg: github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/job
BenchmarkMapResultsToMetricDatas/small_case-10         	    4726	    269036 ns/op	  364367 B/op	    7921 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	    4749	    276557 ns/op	  364366 B/op	    7921 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	    4291	    280005 ns/op	  364369 B/op	    7921 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	    4486	    273217 ns/op	  364368 B/op	    7921 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	    4452	    268510 ns/op	  364368 B/op	    7921 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	    4748	    266801 ns/op	  364367 B/op	    7921 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	      40	  28897098 ns/op	36997460 B/op	  809555 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	      39	  29039734 ns/op	37003479 B/op	  809821 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	      39	  29934966 ns/op	37003487 B/op	  809820 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	      39	  28842040 ns/op	37003451 B/op	  809820 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	      42	  27786820 ns/op	36986299 B/op	  809062 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	      42	  28936529 ns/op	36986283 B/op	  809062 allocs/op
PASS
ok  	github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/job	22.906s
```

pr
```
goos: darwin
goarch: arm64
pkg: github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/job
BenchmarkMapResultsToMetricDatas/small_case-10         	  113601	     10365 ns/op	     444 B/op	       2 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	  117096	     10364 ns/op	     444 B/op	       2 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	  115774	     10357 ns/op	     444 B/op	       2 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	  117190	     10299 ns/op	     445 B/op	       2 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	  116364	     10344 ns/op	     444 B/op	       2 allocs/op
BenchmarkMapResultsToMetricDatas/small_case-10         	  116276	     10241 ns/op	     444 B/op	       2 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	    1191	    955207 ns/op	   15689 B/op	     356 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	    1153	    925903 ns/op	   15933 B/op	     367 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	    1154	    923614 ns/op	   15936 B/op	     367 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	    1190	    928290 ns/op	   15686 B/op	     356 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	    1198	    925005 ns/op	   15635 B/op	     354 allocs/op
BenchmarkMapResultsToMetricDatas/big_case-10           	    1231	    979040 ns/op	   15436 B/op	     345 allocs/op
PASS
ok  	github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/job	50.249s
```